Added obconf

+++ upgrade.sh
@@ -14,7 +14,7 @@ sudo apt-get update
 #sudo apt-get -y upgrade
 sudo apt-get -y dist-upgrade
 sudo apt-get -y install bsdtar links rhino icedtea-plugin rlwrap \
-  tightvncserver xrdp avahi-autoipd espeak gpm
+  tightvncserver xrdp avahi-autoipd espeak gpm obconf
 # Removed: avahi-daemon dosfstools freepats ntfs-3g raspberrypi-net-mods
 # See http://pi.minecraft.net/ & http://www.minecraftforum.net/forum/216-minecraft-pi-edition/
 #rm -rf mcpi
@@ -22,7 +22,7 @@ sudo apt-get -y install bsdtar links rhino icedtea-plugin rlwrap \
 sudo rm -rf ~/mcpi
 rm ~/Desktop/mcpi.desktop
 dirs=$(find /opt/minecraft-pi/ -type d | sed 's|^/opt/minecraft-pi|mcpi|'); for dir in $dirs; do mkdir $dir; done
-dirs=$(find /opt/minecraft-pi/ -type f | sed 's|^/opt/minecraft-pi||'); for dir in $dirs; 
+dirs=$(find /opt/minecraft-pi/ -type f | sed 's|^/opt/minecraft-pi||'); for dir in $dirs;
 do ln -sf '/opt/minecraft-pi'$dir 'mcpi'$dir; done
 
 # Download Python samples
@@ -54,9 +54,11 @@ if [ ! -f ~/Desktop/BYOB.desktop ]; then
 fi
 
 # Download Whiskers, see https://www.raspberrypi.org/forums/viewtopic.php?p=838089
-wget -O Whiskers.image 'https://sites.google.com/site/filedownloader123/files/Scratch.image?attredirects=0&d=1'
-sudo mv Whiskers.image /usr/share/scratch/
-# Start with: scratch.old --image /usr/share/scratch/Whiskers.image
+if [ ! -f /usr/share/scratch/Whiskers.image ]; then
+  wget -O Whiskers.image 'https://sites.google.com/site/filedownloader123/files/Scratch.image?attredirects=0&d=1'
+  sudo mv Whiskers.image /usr/share/scratch/
+  # Start with: scratch.old --image /usr/share/scratch/Whiskers.image
+fi
 
 # See http://www.raspberrypi.org/forums/viewtopic.php?p=534518#p534518
 if [ ! -f ~/abc80sim-2.1/abcdir/masken.bas ]; then
@@ -69,9 +71,9 @@ if [ ! -f ~/abc80sim-2.1/abcdir/masken.bas ]; then
   cd
 fi
 
-# See http://www.raspberrypi.org/forums/viewtopic.php?f=78&t=69420&p=551155
+# See https://www.raspberrypi.org/forums/viewtopic.php?f=78&t=69420&p=775780#p775780
 if [ ! -f /usr/local/bin/x64 ]; then
-  wget -O vice_2.4.20-1_armhf.deb 'https://docs.google.com/uc?authuser=0&id=0B51Q7dpulGC8dEVtTk5ReFBLX0U&export=download'
+  wget -O vice_2.4.20-1_armhf.deb 'https://docs.google.com/uc?id=0B1BUnz1FfQj3bW5FV0hWZnJIcVE&export=download'
   sudo dpkg -i vice_2.4.20-1_armhf.deb
   sudo apt-get -f install
   mkdir ~/c64
@@ -79,6 +81,7 @@ if [ ! -f /usr/local/bin/x64 ]; then
   # x64 -sdlbitdepth 8
   # Save settings within VICE (F12/Settings management/Save current settings), then start with
   # x64
+  # Settings are saved in /home/pi/.vice/sdl-vicerc
   wget 'http://c64.orbin.se/Comal 80 (1985)(Commodore)[a].crt'
   cd
   rm vice_2.4.20-1_armhf.deb
